+++ src/dye/theme.py
@@ -21,10 +21,13 @@
 #
 """classes for storing a theme"""
 
-import benedict
 import jinja2
 import rich
 import tomlkit
+from benedict import benedict
+
+from .exceptions import DyeSyntaxError
+from .utils import benedict_keylist
 
 
 class Theme:
@@ -86,20 +89,14 @@ class Theme:
         env = jinja2.Environment()
 
         try:
-            raw_colors = benedict.benedict(self.definition["colors"])
+            raw_colors = benedict(self.definition["colors"])
         except KeyError:
-            raw_colors = benedict.benedict()
-        self.colors = benedict.benedict()
-
-        # we would use benedict.keypaths() but it sorts the keypaths,
-        # and order is important to us. These two lines of code I stole
-        # out of the soure for keypaths(), they are the ones right before
-        # the .sort()
-        kls = benedict.core.keylists(raw_colors)
-        keylist = [".".join([f"{key}" for key in kl]) for kl in kls]
+            raw_colors = benedict()
+        self.colors = benedict()
 
         # iterate over all the keys in raw_colors, processing and inserting
         # the values into self.colors
+        keylist = benedict_keylist(raw_colors)
         for key in keylist:
             value = raw_colors[key]
             if isinstance(value, str):
@@ -113,8 +110,14 @@ class Theme:
                         colors=self.colors,
                         color=self.colors,
                     )
+            elif isinstance(value, dict):
+                # So we rely on and test for the fact that both the subtable name
+                # (with the dict value) and each of the subtable keys show up in
+                # keylist. Those other keys will be picked up in a different iteration
+                # of the loop. But they shouldn't be a syntax error.
+                pass
             else:
-                self.colors[key] = value
+                raise DyeSyntaxError(f"color {key} must be defined as a string")
 
         # process the elements, using the colors as variables
         # each element in should be a rich.Style() object